+++ sdk/cognitiveservices/Vision.Face/CHANGELOG.md
@@ -1,7 +1,20 @@
 # Release History
 
-## 2.7.0-preview.2 (Unreleased)
+## 2.8.0-preview.2 (Unreleased)
 
+### Features Added
+
+### Breaking Changes
+
+### Key Bugs Fixed
+
+### Fixed
+
+
+## 2.8.0-preview.1 (2021-07-21)
+
+- Add support for the Person Directory feature
+- Note that this version points to the face/v1.0-preview (public preview) endpoints.
 
 ## 2.7.0-preview.1 (2021-03-25)